Preguntar al usuario

+++ install_GoAccess.sh
@@ -1,16 +1,28 @@
 #!/bin/bash
 
-export GOACCESS_VERSION=1.3
+GOACCESS_VERSION=1.3
 
-sudo apt-get install -y libgeoip-dev libncursesw5-dev
+echo "+-------------------------+"
+echo "| GoAccess - Apache Stats |"
+echo "+-------------------------+"
+echo
+echo -n "¿Instalar GoAccess? [s/N] "
+read respuesta
 
-cd
-wget https://tar.goaccess.io/goaccess-$GOACCESS_VERSION.tar.gz
-tar -xzvf goaccess-$GOACCESS_VERSION.tar.gz
-rm goaccess-$GOACCESS_VERSION.tar.gz
+if [ "$respuesta" != "${respuesta#[Ss]}" ] ;then
+    echo Instalando GoAccess...
+    echo
+    sudo apt-get update && sudo apt-get install -y libgeoip-dev libncursesw5-dev
 
-cd goaccess-$GOACCESS_VERSION/
-./configure --enable-utf8 --enable-geoip=legacy
-make
-sudo make install
+    cd
+    wget https://tar.goaccess.io/goaccess-$GOACCESS_VERSION.tar.gz
+    tar -xzvf goaccess-$GOACCESS_VERSION.tar.gz
+    rm goaccess-$GOACCESS_VERSION.tar.gz
 
+    cd goaccess-$GOACCESS_VERSION/
+    ./configure --enable-utf8 --enable-geoip=legacy
+    make
+    sudo make install
+else
+    echo Omitiendo la instalación.
+fi
